feat(fe): make pagination optional
